requirements.txt
aiohttp==3.7.3
aniso8601==9.0.1
appdirs==1.4.4
appnope==0.1.3
astroid==2.4.1
asttokens==2.2.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.7b0
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
comm==0.1.3
Cython==0.29.37
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.3
executing==1.2.0
filelock==3.13.1
Flask==1.1.2
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
flask-praetorian==1.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fsspec==2024.2.0
greenlet==1.1.0
hdbscan==0.8.33
html5lib==1.1
huggingface-hub==0.20.3
idna==2.10
importlib-metadata==6.1.0
inflection==0.3.1
install==1.3.5
ipykernel==6.22.0
ipython==8.11.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==2.11.2
joblib==1.3.2
jupyter_client==8.1.0
jupyter_core==5.3.0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mechanize==0.4.5
mongoengine==0.23.1
MouseInfo==0.1.3
mpmath==1.2.1
multidict==5.1.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.1.0
openai==0.27.2
packaging==23.0
pandas==1.3.0
parso==0.8.3
passlib==1.7.4
pathspec==0.8.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-buzz==1.0.3
PyAutoGUI==0.9.52
pycodestyle==2.7.0
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.1.0
pylint==2.5.2
pymongo==3.12.0
PyMsgBox==1.0.9
pyobjc==7.1
pyobjc-core==7.1
pyobjc-framework-Accounts==7.1
pyobjc-framework-AddressBook==7.1
pyobjc-framework-AdSupport==7.1
pyobjc-framework-AppleScriptKit==7.1
pyobjc-framework-AppleScriptObjC==7.1
pyobjc-framework-ApplicationServices==7.1
pyobjc-framework-Automator==7.1
pyobjc-framework-AVFoundation==7.1
pyobjc-framework-AVKit==7.1
pyobjc-framework-BusinessChat==7.1
pyobjc-framework-CalendarStore==7.1
pyobjc-framework-CFNetwork==7.1
pyobjc-framework-CloudKit==7.1
pyobjc-framework-Cocoa==7.1
pyobjc-framework-Collaboration==7.1
pyobjc-framework-ColorSync==7.1
pyobjc-framework-Contacts==7.1
pyobjc-framework-ContactsUI==7.1
pyobjc-framework-CoreAudio==7.1
pyobjc-framework-CoreAudioKit==7.1
pyobjc-framework-CoreBluetooth==7.1
pyobjc-framework-CoreData==7.1
pyobjc-framework-CoreLocation==7.1
pyobjc-framework-CoreMedia==7.1
pyobjc-framework-CoreMediaIO==7.1
pyobjc-framework-CoreMIDI==7.1
pyobjc-framework-CoreML==7.1
pyobjc-framework-CoreServices==7.1
pyobjc-framework-CoreSpotlight==7.1
pyobjc-framework-CoreText==7.1
pyobjc-framework-CoreWLAN==7.1
pyobjc-framework-CryptoTokenKit==7.1
pyobjc-framework-DictionaryServices==7.1
pyobjc-framework-DiscRecording==7.1
pyobjc-framework-DiscRecordingUI==7.1
pyobjc-framework-DiskArbitration==7.1
pyobjc-framework-DVDPlayback==7.1
pyobjc-framework-EventKit==7.1
pyobjc-framework-ExceptionHandling==7.1
pyobjc-framework-ExternalAccessory==7.1
pyobjc-framework-FinderSync==7.1
pyobjc-framework-FSEvents==7.1
pyobjc-framework-GameCenter==7.1
pyobjc-framework-GameController==7.1
pyobjc-framework-GameKit==7.1
pyobjc-framework-GameplayKit==7.1
pyobjc-framework-ImageCaptureCore==7.1
pyobjc-framework-IMServicePlugIn==7.1
pyobjc-framework-InputMethodKit==7.1
pyobjc-framework-InstallerPlugins==7.1
pyobjc-framework-InstantMessage==7.1
pyobjc-framework-Intents==7.1
pyobjc-framework-IOSurface==7.1
pyobjc-framework-iTunesLibrary==7.1
pyobjc-framework-LatentSemanticMapping==7.1
pyobjc-framework-LaunchServices==7.1
pyobjc-framework-libdispatch==7.1
pyobjc-framework-LocalAuthentication==7.1
pyobjc-framework-MapKit==7.1
pyobjc-framework-MediaAccessibility==7.1
pyobjc-framework-MediaLibrary==7.1
pyobjc-framework-MediaPlayer==7.1
pyobjc-framework-MediaToolbox==7.1
pyobjc-framework-Metal==7.1
pyobjc-framework-MetalKit==7.1
pyobjc-framework-MetalPerformanceShaders==7.1
pyobjc-framework-ModelIO==7.1
pyobjc-framework-MultipeerConnectivity==7.1
pyobjc-framework-NaturalLanguage==7.1
pyobjc-framework-NetFS==7.1
pyobjc-framework-Network==7.1
pyobjc-framework-NetworkExtension==7.1
pyobjc-framework-NotificationCenter==7.1
pyobjc-framework-OpenDirectory==7.1
pyobjc-framework-OSAKit==7.1
pyobjc-framework-Photos==7.1
pyobjc-framework-PhotosUI==7.1
pyobjc-framework-PreferencePanes==7.1
pyobjc-framework-Quartz==7.1
pyobjc-framework-SafariServices==7.1
pyobjc-framework-SceneKit==7.1
pyobjc-framework-ScreenSaver==7.1
pyobjc-framework-ScriptingBridge==7.1
pyobjc-framework-SearchKit==7.1
pyobjc-framework-Security==7.1
pyobjc-framework-SecurityFoundation==7.1
pyobjc-framework-SecurityInterface==7.1
pyobjc-framework-ServiceManagement==7.1
pyobjc-framework-Social==7.1
pyobjc-framework-SpriteKit==7.1
pyobjc-framework-StoreKit==7.1
pyobjc-framework-SyncServices==7.1
pyobjc-framework-SystemConfiguration==7.1
pyobjc-framework-UserNotifications==7.1
pyobjc-framework-VideoSubscriberAccount==7.1
pyobjc-framework-VideoToolbox==7.1
pyobjc-framework-Vision==7.1
pyobjc-framework-WebKit==7.1
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
PyTweening==1.0.3
pytz==2021.1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.0.2
regex==2023.12.25
requests==2.25.1
requests-oauthlib==1.3.0
rubicon-objc==0.4.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
selenium==3.141.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.14.0
slackclient==2.9.3
soupsieve==2.1
SQLAlchemy==1.4.18
stack-data==0.6.2
sympy==1.7.1
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.1
tomli==1.0.4
torch==2.2.0
torchvision==0.10.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.37.2
tweepy==3.10.0
typing_extensions==4.9.0
urllib3==1.26.2
uuid==1.30
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
zipp==3.15.0
